refactor: topic views (#8)
