+++ classes/urls.py
@@ -1,10 +1,10 @@
 from django.urls import path
 
 from .views import (
-    add_class, add_class_next, add_student, delete_class,
-    delete_student, edit_class, edit_class_name, edit_student,
-    list_students, view_passwords,
+    add_class, add_class_next, delete_class,
+    edit_class, edit_class_name, list_students, view_passwords,
 )
+from users.views import edit_student, delete_student, add_student
 
 urlpatterns = [
     path("", list_students),
