+++ classes/views.py
@@ -4,7 +4,7 @@ from django.shortcuts import get_object_or_404, redirect, render
 
 from users.models import InitialPassword
 from classes.models import Stage
-from topic.forms import StageEditForm, StudentForm
+from classes.forms import StageEditForm
 
 # Create your views here.
 context = {}
@@ -52,61 +52,6 @@ def edit_class_name(request, pk):
         return redirect("/your_classes")
     template_name = "form.html"
     context = {"form": form}
-    return render(request, template_name, context)
-
-
-@staff_member_required
-def edit_student(request, pk1, pk2):
-    stage = get_object_or_404(Stage, pk=pk1, teacher=request.user)
-    student = get_object_or_404(User, pk=pk2)
-    form = StudentForm(request.POST or None, instance=student)
-    if form.is_valid():
-        form.save()
-        return redirect("/your_classes")
-    template_name = "form.html"
-    context = {"form": form}
-    return render(request, template_name, context)
-
-
-@staff_member_required
-def delete_student(request, pk1, pk2):
-    stage = get_object_or_404(Stage, pk=pk1, teacher=request.user)
-    student = get_object_or_404(User, pk=pk2)
-    if request.method == "POST":
-        student.delete()
-        return redirect("/your_classes")
-    template_name = "delete_student.html"
-    context = {"stage": stage, "student": student}
-    return render(request, template_name, context)
-
-
-@staff_member_required
-def add_student(request, pk):
-    stage = get_object_or_404(Stage, pk=pk, teacher=request.user)
-    form = StudentForm(request.POST or None)
-    if form.is_valid():
-        student = form.save(commit=False)
-        first_name = student.first_name
-        last_name = student.last_name
-        username = (
-            first_name[: min(3, len(first_name))].strip()
-            + last_name[: min(4, len(last_name))].strip()
-        ).lower()
-        student.username = username
-        counter = 1
-        while User.objects.filter(username=username):
-            username = first_name + str(counter)
-            counter += 1
-        student.save()
-        password = User.objects.make_random_password()
-        student.set_password(password)
-        student.save()
-        stage.students.add(student)
-        stage.save()
-        InitialPassword.objects.create(student=student, password=password)
-        return redirect("/your_classes")
-    context = {"stage": stage, "form": form}
-    template_name = "add_student.html"
     return render(request, template_name, context)
