+++ grade/templates/grade_add.html
@@ -8,18 +8,18 @@
         <a href="{{ skill.get_grade_edit_url }}">Edit latest grades</a><br>
         <a href="{{ skill.get_grade_delete_url }}">Delete latest grades</a></p>
 
-    <form method="POST", action=".">{% csrf_token %}
+    <form method="POST", action="../../topic/templates">{% csrf_token %}
         Add grades below:
         <table>
             <tr>
                 <td>Level:</td>
                 <td>
                     <select name='level'>
-                        <option value=1>Chill</option> 
+                        <option value=1>Chill</option>
                         <option value=2>Medium</option>
-                        <option value=3>Challenge</option>                               
+                        <option value=3>Challenge</option>
                     </select>
-                    
+
                 </td>
             </tr>
 
@@ -31,30 +31,29 @@
                 <td>
                     <li> <label>{{ student.first_name }} {{ student.last_name }}</label></li>
                 </td>
-                
+
                 <td>
                     <select name={{student.pk}}>
-                        <option value="nb">nb</option> 
+                        <option value="nb">nb</option>
                         <option value="tick">✓</option>
                         <option value="cross">☓</option>
                         <option value="G">G</option>
-                        <option value="B">B</option>                               
+                        <option value="B">B</option>
                     </select>
                 </td>
             </tr>
             {% endfor %}
 
 
-            
 
-            
+
+
         </table>
         <br>
         <input type="submit" value="OK"><br><br>
-    </form>   
+    </form>
 
 
 
-    
+
 {% endblock %}
-
