+++ templates/base.html
@@ -9,13 +9,23 @@
     <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-EVSTQN3/azprG1Anm3QDgpJLIm9Nao0Yz1ztcQTwFspd3yD65VohhpuuCOmLASjC" crossorigin="anonymous">
 
     <title>ThinkingFeedback</title>
-            
+
+    <style>
+        table, th, td {
+          border: 1px solid;
+          padding: 15px;
+        }
+        .inner-table {
+            padding: 0;
+        }
+    </style>
+
   </head>
   <body>
     {% include 'navbar.html' %}
     <div class = 'container'>
         <div class = 'row'>
-            <div class = 'col-6 mx-auto'>
+            <div class = 'col-6'>
               <br>
                 <h1>{{ title }}</h1>
                 {% if messages %}
@@ -32,7 +42,7 @@
         </div>
     </div>
 
-    
+
 
     <!-- Optional JavaScript; choose one of the two! -->
