+++ topic/models.py
@@ -48,6 +48,9 @@ class Skill(models.Model):
     def get_grade_delete_url(self):
         return f"/topic/{self.topic.pk}/skill/{self.pk}/grade_delete"
 
+    def get_add_skill_level_url(self):
+        return f"/topic/{self.topic.pk}/skill/{self.pk}/add_skill_level"
+
     def __str__(self):
         return self.title
 
@@ -59,9 +62,19 @@ class SkillLevel(models.Model):
         choices=(("1", "Chill"), ("2", "Medium"), ("3", "Challenge")),
         default=1,
     )
-    example_task = models.ForeignKey(Task, default=1, on_delete=models.CASCADE)
+    description = models.CharField(max_length=200)
+    example_task = models.ForeignKey(Task, default=None, null=True, on_delete=models.CASCADE)
     tasks = models.ManyToManyField(Task, related_name="skill_levels")
     generators = models.ManyToManyField(TaskGenerator, related_name="skill_levels")
+
+    def get_add_skill_to_skill_level_url(self):
+        return f"/topic/{self.skills.all()[0].topic.pk}/skill_level/{self.pk}/add_skill_to_skill_level"
+
+    def get_edit_url(self):
+        return f"/topic/{self.skills.all()[0].topic.pk}/skill_level/{self.pk}/edit"
+
+    def get_delete_url(self):
+        return f"/topic/{self.skills.all()[0].topic.pk}/skill_level/{self.pk}/delete"
 
 
 # What happens here??
