+++ topic/views.py
@@ -1,14 +1,13 @@
 from django.contrib.admin.views.decorators import staff_member_required
-from django.contrib.auth.models import User
+from django.forms import ModelMultipleChoiceField
 from django.shortcuts import get_object_or_404, redirect, render
 
 # Create your views here.
-from topic.models import Skill, Topic
-from grade.models import Grade
+from topic.models import Skill, Topic, SkillLevel
 from classes.models import Stage
 
 from .forms import (
-    SkillModelForm, TopicModelForm,
+    SkillModelForm, TopicModelForm, SkillLevelModelForm, SkillLevelNewSkillModelForm,
 )
 
 # from django.contrib.auth.forms import UserCreationForm, PasswordChangeForm, PasswordResetForm
@@ -18,7 +17,7 @@ context = {}
 
 
 @staff_member_required
-def by_subject(request):
+def stage_all_topics(request):
     stages = Stage.objects.filter(teacher=request.user)
     context["stages"] = stages
 
@@ -31,20 +30,28 @@ def by_subject(request):
     if len(stages) == 0:
         context["empty"] = 1
 
-    template_name = "by_subject.html"
+    template_name = "stage_all_topics.html"
     return render(request, template_name, context)
 
 
 @staff_member_required
 def topic_detail_view(request, pk):
     stages = Stage.objects.filter(teacher=request.user)
-    obj = get_object_or_404(Topic, pk=pk, stage__in=stages)
+    topic = get_object_or_404(Topic, pk=pk, stage__in=stages)
 
-    qs2 = Skill.objects.filter(topic=obj)
+    skill_objects = Skill.objects.filter(topic=topic)
+    skills = []
+    for skill in skill_objects:
+        levels = {
+            "chill": skill.levels.filter(level__startswith="1").all(),
+            "medium": skill.levels.filter(level__startswith="2").all(),
+            "challenge": skill.levels.filter(level__startswith="3").all(),
+        }
+        skills.append((skill, levels))
     template_name = "topic_detail.html"
-    context = {"object": obj, "skill_list": qs2}
+    context = {"topic": topic, "skills": skills}
     context["empty"] = 0
-    if len(qs2) == 0:
+    if len(skills) == 0:
         context["empty"] = 1
     return render(request, template_name, context)
 
@@ -54,11 +61,9 @@ def add_topic(request):
     user = request.user
     form = TopicModelForm(user, request.POST or None)
     if form.is_valid():
-        obj = form.save(commit=False)
-        obj.teacher = request.user
-        obj.save()
-        user = request.user
-        form = TopicModelForm(user)
+        topic = form.save(commit=False)
+        topic.teacher = request.user
+        topic.save()
         return redirect("/topic")
     template_name = "form.html"
     context = {"form": form}
@@ -70,11 +75,42 @@ def add_skill(request, pk):
     stages = Stage.objects.filter(teacher=request.user)
     form = SkillModelForm(request.POST or None)
     if form.is_valid():
-        obj = form.save(commit=False)
+        skill = form.save(commit=False)
         topic = get_object_or_404(Topic, pk=pk, stage__in=stages)
-        obj.topic = topic
-        obj.save()
+        skill.topic = topic
+        skill.save()
         return redirect("..")
+    template_name = "form.html"
+    context = {"form": form}
+    return render(request, template_name, context)
+
+
+@staff_member_required
+def add_skill_level(request, pk1, pk2):
+    form = SkillLevelModelForm(request.POST or None)
+    if form.is_valid():
+        skill_level = form.save(commit=False)
+        skill = get_object_or_404(Skill, pk=pk2)
+        skill_level.save()
+        skill_level.skills.add(skill)
+        skill_level.save()
+        return redirect("../../..")
+    template_name = "form.html"
+    context = {"form": form}
+    return render(request, template_name, context)
+
+
+@staff_member_required
+def add_skill_to_skill_level(request, pk1, pk3):
+    topic = get_object_or_404(Topic, pk=pk1)
+    skill_level = get_object_or_404(SkillLevel, pk=pk3)
+    form = SkillLevelNewSkillModelForm(topic, request.POST or None, instance=skill_level)
+    if form.is_valid():
+        skill_level = get_object_or_404(SkillLevel, pk=pk3)
+        for skill in form.cleaned_data["skills"]:
+            skill_level.skills.add(skill)
+        skill_level.save()
+        return redirect("../../..")
     template_name = "form.html"
     context = {"form": form}
     return render(request, template_name, context)
@@ -111,146 +147,7 @@ def delete_topic(request, pk):
 
 
 @staff_member_required
-def skill_update(request, pk1, pk2):
-    stages = Stage.objects.filter(teacher=request.user)
-    topic = get_object_or_404(Topic, pk=pk1, stage__in=stages)
-    skill = get_object_or_404(Skill, pk=pk2)
-    stage = topic.stage
-    template_name = "skill_update.html"
-    students = stage.students.all()
-    context = {"topic": topic, "skill": skill, "stage": stage, "students": students}
-
-    if request.POST:
-        for key in request.POST.keys():
-            if key != "csrfmiddlewaretoken" and key != "level":
-                student = User.objects.get(pk=key)
-                value = request.POST[key]
-                level = request.POST["level"]
-                grade = Grade(student=student, skill=skill, value=value, level=level)
-                grade.save()
-        return redirect("../..")
-    return render(request, template_name, context)
-
-
-@staff_member_required
-def skill_grade_edit(request, pk1, pk2):
-    stages = Stage.objects.filter(teacher=request.user)
-    topic = get_object_or_404(Topic, pk=pk1, stage__in=stages)
-    skill = get_object_or_404(Skill, pk=pk2)
-    context = {"topic": topic, "skill": skill, "title": "Edit grades"}
-    template_name = "skill_grade_edit.html"
-
-    if request.POST:
-        level = request.POST["level"]
-        return redirect(f"./{level}/")
-    return render(request, template_name, context)
-
-
-@staff_member_required
-def skill_grade_edit_next(request, pk1, pk2, level):
-    stages = Stage.objects.filter(teacher=request.user)
-    topic = get_object_or_404(Topic, pk=pk1, stage__in=stages)
-    skill = get_object_or_404(Skill, pk=pk2)
-    context = {"topic": topic, "skill": skill, "level": "Chill", "title": "Edit grades"}
-    if int(level) == 2:
-        context["level"] = "Medium"
-    if int(level) == 3:
-        context["level"] = "Challenge"
-
-    stage = topic.stage
-    students = stage.students.all()
-    add = {"stage": stage, "students": students}
-    context = {**context, **add}
-
-    last_grades = {}
-    for student in students:
-        grades = Grade.objects.filter(student=student, skill=skill, level=level)
-        if len(grades) != 0:
-            last_grades[student] = grades.last()
-    context["last_grades"] = last_grades
-
-    if len(last_grades) == 0:
-        return render(
-            request,
-            "base.html",
-            {
-                "content": "Oh, there seems that there are no grades within this skill and level!",
-            },
-        )
-
-    if request.POST:
-        for key, val in request.POST.items():
-            if key != "csrfmiddlewaretoken":
-                student = get_object_or_404(User, pk=key)
-                if val == "empty":
-                    last_grades[student].delete()
-                elif val != last_grades[student].value:
-                    last_grades[student].value = val
-                    last_grades[student].save()
-
-        return redirect("../..")
-
-    template_name = "skill_grade_edit_next.html"
-    return render(request, template_name, context)
-
-
-@staff_member_required
-def skill_grade_delete(request, pk1, pk2):
-    stages = Stage.objects.filter(teacher=request.user)
-    topic = get_object_or_404(Topic, pk=pk1, stage__in=stages)
-    skill = get_object_or_404(Skill, pk=pk2)
-    context = {"topic": topic, "skill": skill, "title": "Delete grades"}
-    template_name = "skill_grade_edit.html"
-
-    if request.POST:
-        level = request.POST["level"]
-        return redirect(f"./{level}/")
-    return render(request, template_name, context)
-
-
-@staff_member_required
-def skill_grade_delete_next(request, pk1, pk2, level):
-    stages = Stage.objects.filter(teacher=request.user)
-    topic = get_object_or_404(Topic, pk=pk1, stage__in=stages)
-    skill = get_object_or_404(Skill, pk=pk2)
-    context = {"topic": topic, "skill": skill, "level": "Chill"}
-    if int(level) == 2:
-        context["level"] = "Medium"
-    if int(level) == 3:
-        context["level"] = "Challenge"
-
-    stage = topic.stage
-    students = stage.students.all()
-    add = {"stage": stage, "students": students}
-    context = {**context, **add}
-
-    last_grades = {}
-    for student in students:
-        grades = Grade.objects.filter(student=student, skill=skill, level=level)
-        if len(grades) != 0:
-            last_grades[student] = grades.last()
-    context["last_grades"] = last_grades
-
-    if len(last_grades) == 0:
-        return render(
-            request,
-            "base.html",
-            {
-                "content": "Oh, there seems that there are no grades within this skill and level!",
-            },
-        )
-
-    if request.POST:
-        for grade in last_grades.values():
-            grade.delete()
-        return redirect("../..")
-
-    template_name = "skill_grade_delete_next.html"
-    return render(request, template_name, context)
-
-
-@staff_member_required
-def skill_edit(request, pk1, pk2):
+def update_skill(request, pk1, pk2):
     stages = Stage.objects.filter(teacher=request.user)
     topic = get_object_or_404(Topic, pk=pk1, stage__in=stages)
     skill = get_object_or_404(Skill, pk=pk2)
@@ -259,20 +156,45 @@ def skill_edit(request, pk1, pk2):
         obj = form.save(commit=False)
         obj.topic = topic
         obj.save()
-        return redirect(f"../../{obj.pk}")
+        return redirect(f"../../../")
     template_name = "form.html"
     context = {"form": form}
     return render(request, template_name, context)
 
 
 @staff_member_required
-def skill_delete(request, pk1, pk2):
+def delete_skill(request, pk1, pk2):
     stages = Stage.objects.filter(teacher=request.user)
     topic = get_object_or_404(Topic, pk=pk1, stage__in=stages)
     skill = get_object_or_404(Skill, pk=pk2)
     if request.method == "POST":
         skill.delete()
-        return redirect("../../..")
+        return redirect("../../../")
     template_name = "delete_skill.html"
     context = {"skill": skill}
+    return render(request, template_name, context)
+
+
+@staff_member_required
+def update_skill_level(request, pk1, pk3):
+    skill_level = get_object_or_404(SkillLevel, pk=pk3)
+    form = SkillLevelModelForm(request.POST or None, instance=skill_level)
+    if form.is_valid():
+        skill_level_edit = form.save(commit=False)
+        # skill_level_edit.skills = skill_level.skills
+        skill_level_edit.save()
+        return redirect(f"../../../")
+    template_name = "form.html"
+    context = {"form": form}
+    return render(request, template_name, context)
+
+
+@staff_member_required
+def delete_skill_level(request, pk1, pk3):
+    skill_level = get_object_or_404(SkillLevel, pk=pk3)
+    if request.method == "POST":
+        skill_level.delete()
+        return redirect("../../../")
+    template_name = "delete_skill.html"
+    context = {"skill_level": skill_level}
     return render(request, template_name, context)
